% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliment_function.R
\name{compliment_me}
\alias{compliment_me}
\title{A Compliment Function}
\usage{
compliment_me(x)
}
\arguments{
\item{x}{Sometimes you need to be reminded how awesome you are.  Type "please" or it will give you sass.}
}
\description{
This function gives you a compliment.
}
\examples{
compliment_me()
}
\keyword{compliment}
