% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eightball_function.R
\name{magic8ball}
\alias{magic8ball}
\title{Magic Eight Ball Function}
\usage{
magic8ball(x = TRUE)
}
\arguments{
\item{x}{Do you have an imperative yes or no question?  Ask the eight ball.  Defaults to "TRUE."}
}
\description{
This function gives you one of the outcomes of the Magic Eight Ball.
}
\examples{
magic8ball()
}
\keyword{ball,}
\keyword{eight}
\keyword{magic}
